pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dragonfly-pulse"
version = "0.1.0"
description = "Dragonfly Pulse: 校园科技赛雷达与竞赛情报系统"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Thedan-1" }
]
keywords = ["crawler", "competition", "education", "llm", "china"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Intended Audience :: Education",
]
dependencies = [
  "requests>=2.32.0",
  "beautifulsoup4>=4.12.0",
  "trafilatura>=1.9.0",
  "readability-lxml>=0.8.1",
  "python-dateutil>=2.9.0",
  "dateparser>=1.2.0",
  "SQLAlchemy>=2.0.30",
  "PyYAML>=6.0.1",
  "APScheduler>=3.10.4",
  "python-dotenv>=1.0.1",
  "pytest>=8.3.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "mcp>=1.2.0",
]

[project.urls]
Homepage = "https://github.com/Thedan-1/dragonfly-pulse"
Repository = "https://github.com/Thedan-1/dragonfly-pulse"
Issues = "https://github.com/Thedan-1/dragonfly-pulse/issues"

[project.scripts]
dragonfly-pulse-init-db = "scripts.init_db:main"
dragonfly-pulse-run-daily = "scripts.run_daily:main"
dragonfly-pulse-run-api = "scripts.run_api:main"
dragonfly-pulse-run-mcp = "scripts.run_mcp:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
